[python] Factor memory utility functions in one module
