+++ bindings/pyroot/pythonizations/CMakeLists.txt
@@ -84,6 +84,7 @@ set(py_sources
   ROOT/_pythonization/_cppinstance.py
   ROOT/_pythonization/_drawables.py
   ROOT/_pythonization/_generic.py
+  ROOT/_pythonization/_memory_utils.py
   ROOT/_pythonization/_pyz_utils.py
   ROOT/_pythonization/_runtime_error.py
   ROOT/_pythonization/_rvec.py
