+++ bindings/pyroot/pythonizations/python/ROOT/_pythonization/_th1.py
@@ -9,7 +9,7 @@
 ################################################################################
 
 from . import pythonization
-
+from ROOT._pythonization._memory_utils import _constructor_releasing_ownership, _SetDirectory_SetOwnership
 
 # Multiplication by constant
 
@@ -21,30 +21,6 @@ def _imul(self, c):
     # - A multiplied histogram (in place)
     self.Scale(c)
     return self
-
-
-def _should_give_up_ownership(object):
-    """
-    Ownership of objects which automatically register to a directory should be
-    left to C++, except if the object is gROOT.
-    """
-    import ROOT
-    tdir = object.GetDirectory()
-    return bool(tdir) and tdir is not ROOT.gROOT
-
-
-def _constructor_releasing_ownership(self, *args, **kwargs):
-    """
-    Forward the arguments to the C++ constructor and give up ownership if the
-    object is attached to a directory, which is then the owner. The only
-    exception is when the owner is gROOT, to avoid introducing a
-    backwards-incompatible change.
-    """
-    import ROOT
-
-    self._cpp_constructor(*args, **kwargs)
-    if _should_give_up_ownership(self):
-        ROOT.SetOwnership(self, False)
 
 
 def inject_constructor_releasing_ownership(klass):
@@ -67,12 +43,6 @@ _th1_derived_classes_to_pythonize = [
 for klass in _th1_derived_classes_to_pythonize:
     pythonization(klass)(inject_constructor_releasing_ownership)
 
-def _SetDirectory_SetOwnership(self, dir):
-    self._Original_SetDirectory(dir)
-    if dir:
-        # If we are actually registering with a directory, give ownership to C++
-        import ROOT
-        ROOT.SetOwnership(self, False)
 
 @pythonization('TH1')
 def pythonize_th1(klass):
